issue #663: upgraded web3sdk to java-sdk, apapt java-sdk.  (#666)
* fix bug: the monitor bash script in weevent-broker execute failed

* optimize client retry when upload failed

* optimize code

* issue #466: add common code/messsage for weevent-broker's restful response

* revert jsonRpc response

* optimize code format

* reset code

* issue #505: replace jsonrpc with rest for weevent-client

* remove timeout for jsonrpc

* resert code

* reset code

* optimize client

* optimize WeEventClient code

* add brokerRestfulUrl

* remove unused ERROR_CODE

* issue #545: add test for STOMP Protocol

* add jackson version

* optimize code

* issue #457: separate jms from weevent-client and provide a weevent-jms.jar

* remove defaultBrokerUrl

* remove unused @Slf4j

* fix ci issue

* issue #458: provide an asynchronous publish interface in weevent-client

* optimize code

* optimize stompsend

* add weevent-jms in settings.gradle

* optimize stompsend

* merge upstream

* test ci

* optimize timeout

* add jboss

* test ci

* remove unused timeout

* add comment for repository jboss

* optimize jboss repository

* issue #563: integrate large file transfer function into weevent-governance

* remove unused property

* optimize code

* optimize deploy topic bash script

* add copy fisco credentials when execute install-governance.sh

* add test for FileController in weevent-governance

* copy fisco credentials

* add test for FileController

* add test for FileController

* optimize JsonHelper doc

* optimize code

* revert utils

* test ci

* test ci

* test ci

* test ci

* test ci

* test ci

* test ci

* test ci

* test ci

* test ci

* optimize code

* revert ci bash script

* revise subscribe interface for weevent-client

* optimize code

* fix download file Chinese name garbled

* fix bug: fileName is null when restart weevent-broker

* add check file is uploaded

* fix bug: add check file isUploaded, add table for file transport status

* optimize code

* optimize code

* format code

* fix bug: fix zk start failed bug

* upgarde web3sdk to 2.2.5

* optimize sql

* fix bug: check whether the npm command is installed before executing the build-web.sh

* fix bug: fix the problem of indirect dependence between subprojects

* fix bug: add open topic before publish/subscribe

* add check groupId exist when build IWeEventClient

* optimize validateGroupId

* optimize comment

* optimize code

* upgrade zookeeper and jackson

* adapt BcosSDK

* adapt BcosSDK

* adapt bcosSDK

* optimize amop code

* adapt amop

* remove unused code

* remove unused code

* remove unused code

+++ weevent-core/src/main/java/com/webank/weevent/core/fisco/FiscoBcosBroker4Producer.java
@@ -41,15 +41,15 @@ public class FiscoBcosBroker4Producer extends FiscoBcosTopicAdmin implements IPr
         // publishEvent support async operator in callback
         if (event.getExtensions().containsKey(WeEvent.WeEvent_SIGN)) {
             return fiscoBcosDelegate.sendRawTransaction(event.getTopic(),
-                    Long.parseLong(groupId),
+                    Integer.parseInt(groupId),
                     new String(event.getContent(), StandardCharsets.UTF_8));
         } else {
             if (event.getExtensions().containsKey(WeEvent.WeEvent_EPHEMERAL)) {
                 log.info("ephemeral event");
-                return fiscoBcosDelegate.sendAMOP(event.getTopic(), Long.parseLong(groupId), JsonHelper.object2Json(event));
+                return fiscoBcosDelegate.sendAMOP(event.getTopic(), Integer.parseInt(groupId), JsonHelper.object2Json(event));
             } else {
                 return fiscoBcosDelegate.publishEvent(event.getTopic(),
-                        Long.parseLong(groupId),
+                        Integer.parseInt(groupId),
                         new String(event.getContent(), StandardCharsets.UTF_8),
                         JsonHelper.object2Json(event.getExtensions()));
             }
